Fixed one of the mem issues potentially. Caused by race condition with DOTS deallocating before on destroy is called

+++ Assets/_Project/Packages/Grain Synth/Scripts/Authoring/BaseEmitterClass.cs
@@ -76,7 +76,7 @@ public class BaseEmitterClass : MonoBehaviour, IConvertGameObjectToEntity
 
     public void DestroyEntity()
     {
-        if (_EmitterEntity != null)
+        if (World.All.Count != 0 && _EmitterEntity != null)
             _EntityManager.DestroyEntity(_EmitterEntity);
     }
